[FIX] Path of moved imports
